Add support for 'sprite-scale' directive
The new sprite-scale directive applies to an entire sprite, and tells
the tool that the intended usage is for high DPI devices with the
specified pixel ratio. Normally, this value is "1"; for devices like the
Retina display, the recommended value is "2". Thus, any image you add to
a sprite with sprite-scale: 2; should have double the dimensions.

The sprite-scale directive takes a float as input. If the scale is not
1, it emits an additional background-size tag for each sprite reference,
and automatically adjusts the background-position based on the scale
ratio.

If the tool detects the scale ratio will result in fractional values
(usually an indication of an error), it will emit a WARNING.